protocol/workspace: Move cosmic dispatch code to seperate file
